fix: refactor thundercloud, add notification.js
